Removing unnecessary depdencies (#13)
This pull request includes updates to dependencies and logging
configuration, focusing on simplifying the codebase and improving
readability. The most important changes involve the removal of unused
dependencies and the modification of the logger configuration to ensure
consistent pretty-printing.

### Dependency Updates:
*
[`package.json`](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L28-L44):
Removed unused dependencies such as `@types/express`, `express`,
`nodemon`, `ts-node`, and `npm-run-all` to streamline the project.
[[1]](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L28-L44)
[[2]](diffhunk://#diff-7ae45ad102eab3b6d7e7896acd08c427a9b25b346470d7bc6507b6481575d519L65)

### Logging Configuration:
*
[`src/utils/logger.ts`](diffhunk://#diff-0b242f99cdb5b36e8a0aa886862c30c2f480aca7512bcfc3e2fa4889055d4544L11-R21):
Updated the logger configuration to always use pretty-printing for
consistent, readable output, regardless of the environment. Simplified
the `transport` property by removing conditional logic.